Refactor: useAuth isLogin 계산 로직 useMemo로 수정
