+++ src/components/layout/authActions.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useAuth } from "@/hooks/common/useAuth";
+import { useAuth } from "@/hooks/auth/useAuth";
 import Link from "next/link";
 
 export default function AuthActions() {
