+++ src/components/nav/navMenu.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useAuth } from "@/hooks/common/useAuth";
+import { useAuth } from "@/hooks/auth/useAuth";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
