+++ src/hooks/auth/useAuth.ts
@@ -1,18 +1,13 @@
 import { useAuthStore } from "@/store/useAuthStore";
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
+import { useMemo } from "react";
 
 export function useAuth() {
   const router = useRouter();
   const { token, clearToken, _hasHydrated } = useAuthStore();
-  const [isLogin, setIsLogin] = useState(false);
 
-  useEffect(() => {
-    if (_hasHydrated) {
-      setIsLogin(!!token);
-    } else {
-      setIsLogin(false);
-    }
+  const isLogin = useMemo(() => {
+    return _hasHydrated && !!token;
   }, [_hasHydrated, token]);
 
   // 로그아웃
